+++ src/layouts/index.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Header from '../components/Header'
-import Footer from '../components/Footer';
+//import Footer from '../components/Footer';
 import '../styles/app.scss'
 
 const TemplateWrapper = ({ children }) => (
   <div>
     <Header />
     {children()}
-    <Footer />
+    {/*<Footer />*/}
   </div>
 )
