--- src/pages/empresa/index.md
@@ -1,14 +1,0 @@
----
-templateKey: 'empresa-page'
-path: /empresa
-title: Quem somos
-heading: Quanto maior o <strong>conhecimento sobre seu cliente</strong>, maior o valor que você será capaz de entregar
-description:
-    <p>
-    A Cinnecta é uma empresa de tecnologia especializada em Gerenciamento de Comportamento do Usuário para canais digitais móveis. Fornecemos soluções inovadoras que permitem às empresas ter visibilidade sobre o Perfil de Uso e Experiência do seus clientes por meio do monitoramento do uso de smartphones, proporcionando o aumento do Engajamento e da Retenção.
-    </p>
-
-    <p>
-    Entendemos que a forma como as pessoas interagem com seu smartphone revela seus interesses, poder aquisitivo e perfil de consumo. Ter visibilidade sobre esses aspectos é um dos grandes desafios para empresas. A Cinnecta fornece soluções que permitem conhecer profundamente o cliente através da analise de dados de comportamento e de geolocalização.
-    </p>
----
