--- src/pages/index.md
@@ -1,22 +1,0 @@
----
-templateKey: 'index-page'
-path: /
-title: Home
-
-banner: 
-    heading: Conheça seus clientes.<br />Fidelize sua marca.
-    subheading: Entenda o comportamento do seu cliente e gere comunicação personalizada, assertiva e eficaz.
-    callToAction: Conheça nossas soluções
-
-featured:
-    heading: Entenda o comportamento do seu cliente e aumente o engajamento nos canais digitais móveis
-    callToAction: Solicite o contato de um consultor
-    reactive:
-        heading: 'PROCESSOS REATIVOS de análise de comportamento do usuário:'
-        content: 
-            Ter a visibilidade do comportamento dos usuários do seu app é um dos maiores desafios atuais das empresas. Analisar o comportamento e segmentar a base de usuários de acordo com as experiências e o perfil de consumo de cada um é uma necessidade para permitir comunicação mais assertiva e garantir maior engajamento com sua marca. Processos tradicionais de identificação de comportamento e coleta de feedback de clientes são reativos, feitos por amostragem e apresentam resultado geralmente impreciso.
-    proactive:
-        heading: 'PROCESSOS PROATIVOS usando as soluções da Cinnecta:'
-        content:
-            Com a solução da CINNECTA sua empresa terá a visibilidade do comportamento e da rotina de cada usuário. Isso permite o planejamento e geração de campanhas individualizadas e proativas para aumento de engajamento e retenção no seu app. Integrando nossa plataforma, você terá acesso a dados valiosos sobre o perfil de cada usuário e, através de uma interface gráfica amigável e intuitiva, poderá planejar as ações para minimizar o impacto de experiências negativas, investidas da concorrência ou para aumentar o engajamento.
----
