--- src/pages/tecnologia.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const TecnologiaPage = () => (
-    <div>
-        Página de Tecnologia
-    </div>
-)
-
-export default TecnologiaPage;
