+++ src/pages_old/tecnologia.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const TecnologiaPage = () => (
+    <div>
+        Página de Tecnologia
+    </div>
+)
+
+export default TecnologiaPage;
