+++ src/styles/_settings.scss
@@ -63,6 +63,8 @@
 @import "~foundation-sites/scss/util/util";
 @import url('https://fonts.googleapis.com/css?family=Barlow:400,500,600,700');
 @import url('https://file.myfontastic.com/ZUCYuU4b9PQafWCX9diMME/icons.css');
+@import "./fonts/fonts.css";
+
 
 // 1. Global
 // ---------
@@ -71,8 +73,8 @@ $global-font-size: 100%;
 $global-width: rem-calc(1200);
 $global-lineheight: 1.5;
 $foundation-palette: (
-  primary: #304972,
-  secondary: #FF6864,
+  primary: #0055ff,
+  secondary: #00ddee,
   success: #3adb76,
   warning: #ffae00,
   alert: #cc4b37,
@@ -86,7 +88,7 @@ $black: #0a0a0a;
 $white: #fefefe;
 $body-background: $white;
 $body-font-color: $black;
-$body-font-family: Barlow, 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif;
+$body-font-family: 'Proxima Nova', 'Proxima Nova Alt', 'proxima-nova', 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif;
 $body-antialiased: true;
 $global-margin: 1rem;
 $global-padding: 1rem;
